changed (unsigned) and (unsigned int) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11750 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Groups/ORGroup.h
@@ -33,11 +33,11 @@
 #pragma mark ¥¥¥Assessors
 - (void) setOrcaObjects:(NSMutableArray*)anArray;
 - (NSMutableArray*) orcaObjects;
-- (unsigned int) count;
+- (NSUInteger) count;
 - ( NSMutableArray*)children;
 - (NSMutableArray*) familyList;
 - (NSEnumerator*) objectEnumerator;
-- (id) objectAtIndex:(unsigned) index;
+- (id) objectAtIndex:(NSUInteger) index;
 
 #pragma mark ¥¥¥Undoable Actions
 - (void) addObject:(id)anObject;
